refactor: :recycle: refactor Events to use enum and better type checking
There was a problem with the onHover event that led me down a rabbit hole. With multiple developers it seems almost a necessity to use an enum for events to make sure we aren't doing the same thing with different string names. I also added better typing for the fireEvent method to make sure you are passing the correct data. Sorry this is a large change.